+++ CHANGELOG.md
@@ -40,6 +40,15 @@
 - **settings**: Remove USE_L10N and SHA1PasswordHasher as they are no longer in use
 - **tests**: Remove USE_L10N as it no longer in use.
 
+
+## 6.4.1 (2025-10-13)
+
+### Fix
+
+- Remove RQEmailBackend
+- **devilry_message**: Translation and logging fix.
+- **message**: Make devilry_message send via RQ - thus deprecating RQEmailBackend.
+
 ## 6.4.0 (2025-04-11)
 
 ## 6.4.0rc5 (2025-04-10)